+++ src/tests/insert/split-root.cpp
@@ -31,9 +31,9 @@ bool split_root(
 	int RDMA_TYPE,
 	int exec,
 
-	uint8_t *hbm,
-	uint8_t *req_buffer,
-	uint8_t *resp_buffer
+	uint64_t *hbm,
+	uint64_t *req_buffer,
+	uint64_t *resp_buffer
 ) {
 	bool pass = true;
 	bptr_t root = 0;
